Refactor CLI into its own file.
