+++ cumulusci/tasks/bulkdata/data_generation/data_generator.py
@@ -14,7 +14,7 @@ from .data_generator_runtime import output_batches
 # 2 a) merge options informtion from the parse with options from the
 #      environment
 # 3. Generating the objects top to bottom (including evaluating Jinja) in
-#    generate_from_yaml.output_batches
+#    data_generator_runtime.output_batches
 #
 # The function generate at the bottom of this file is the entry point to all
 # of it.
@@ -90,3 +90,9 @@ def generate(open_yaml_file, count=1, cli_options=None, output_stream=None):
     )
 
     return ExecutionSummary(parse_result, runtime_context)
+
+
+if __name__ == "__main__":
+    from .data_generator_cli import generate_cli
+
+    generate_cli()
